Added release source (0.2.4-beta)
The source code is provided as-is and has not been modified from the
build in the Chrome Webstore.
As I don’t have anymore the time to work on cardboard, it’s better for
everyone
that I open source it rather than seeing it not evolving and getting
outdated.
At the time I started coding cardboard, I did not know about a lot of
things (it was my first Angular project.)
Now it can be improved in a lot of way. Some ideas:
- use npm
- use bower
- create tests
- update angular
- load headers from external source (then cache it) to reduce the
extension footprint
- I'm sure you guys will come with a lot of cool stuff to breathe new
life into this project ;)