2.0 rewrite
Almost every line of code has been rewritten
Hopefully my experience in Angular, front-end and JS development has
improved enough to make cardboard better overall.
Of course it's not over yet. And I'll keep updating this repo as I
improve cardboard.
In bulk:
- Angular 1.3
- Material Design
- Promises (bluebird)
- Modulable code
- ground-work for moveable cards (packery)
- permissions and settings managed way better (under the hood)
- Google trends
- Removed search bar (redundant with omnibar)
- Some features missing (like app ordered by frequency and few others)
but will be reintroduced later, for CWS publishing.
- New favicon source (google)
- Bookmark card now shows all bookmarks (not only recents)
- Permissions are intelligently requested (no more "grant" button)

This is just the beginning of v2.0  ;)