New Fab behaviour + moved cards to feedCtrl
Before:
on mouseenter the fab would show its children and on mouseleave it
would make them transparent.
This was not ideal because the tooltips would show up and by hovering
on the empty area over the fab it would bring back the children.

Now:
fab requires a mouse click to show its children.
on mouseleave the children are display: none which means they don't
block what's behind anymore. (animation are preserved)

So there is a bit of jQuery code but couldn't go without it with the
materializecss framework.

Fab and Cards stuff are moved to the FeedCtrl for a more modulable
code. (ligthen the rootctrl)